[ops] move !extract-* to ops/extract.py
